requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/base.txt
    #   kombu
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
billiard==4.2.1
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.5.1
    # via -r requirements/base.txt
certifi==2025.1.31
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==2.3.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
    #   edx-toggles
coverage[toml]==7.8.0
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-mock-queries
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   edx-toggles
    #   model-bakery
    #   openedx-filters
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-toggles
django-mock-queries==2.3.0
    # via -r requirements/test.in
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-toggles
djangorestframework==3.16.0
    # via
    #   -r requirements/base.txt
    #   django-mock-queries
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-django-utils==7.4.0
    # via
    #   -r requirements/base.txt
    #   edx-toggles
edx-opaque-keys[django]==2.12.0
    # via
    #   -r requirements/base.txt
    #   openedx-filters
edx-toggles==5.3.0
    # via -r requirements/base.txt
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   xblock
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   -r requirements/base.txt
    #   code-annotations
kombu==5.5.3
    # via
    #   -r requirements/base.txt
    #   celery
lxml==5.3.2
    # via
    #   -r requirements/base.txt
    #   xblock
mako==1.3.10
    # via
    #   -r requirements/base.txt
    #   xblock
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
model-bakery==1.20.4
    # via django-mock-queries
newrelic==10.9.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
openedx-atlas==0.7.0
    # via -r requirements/base.txt
openedx-filters==2.0.1
    # via -r requirements/base.txt
packaging==25.0
    # via pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
prison==0.2.1
    # via -r requirements/base.txt
prompt-toolkit==3.0.51
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pymongo==4.12.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
    #   xblock
python-slugify==8.0.4
    # via
    #   -r requirements/base.txt
    #   code-annotations
pytz==2025.2
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   responses
    #   superset-api-client
    #   xblock
redis==5.2.1
    # via -r requirements/base.txt
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   responses
    #   superset-api-client
requests-oauthlib==2.0.0
    # via
    #   -r requirements/base.txt
    #   superset-api-client
responses==0.25.7
    # via -r requirements/test.in
simplejson==3.20.1
    # via
    #   -r requirements/base.txt
    #   xblock
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   prison
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
superset-api-client==0.6.0
    # via -r requirements/base.txt
text-unidecode==1.3
    # via
    #   -r requirements/base.txt
    #   python-slugify
typing-extensions==4.13.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
tzdata==2025.2
    # via
    #   -r requirements/base.txt
    #   kombu
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   requests
    #   responses
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
web-fragments==3.0.0
    # via
    #   -r requirements/base.txt
    #   xblock
webob==1.8.9
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==5.2.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
